moved initialization of unleash client from init file
