+++ koku/hcs/tasks.py
@@ -8,8 +8,8 @@ import logging
 from api.common import log_json
 from api.provider.models import Provider
 from api.utils import DateHelper
-from hcs import enable_HCS_processing
 from koku import celery_app
+from koku.feature_flags import UNLEASH_CLIENT
 
 
 LOG = logging.getLogger(__name__)
@@ -18,6 +18,16 @@ HCS_QUEUE = "hcs"
 
 # any additional queues should be added to this list
 QUEUE_LIST = [HCS_QUEUE]
+
+
+def enable_HCS_processing(source_uuid, source_type, account):  # noqa
+    """Helper to determine if source is enabled for HCS."""
+    if account and not account.startswith("acct"):
+        account = f"acct{account}"
+
+    context = {"schema": account}
+    LOG.info(f"enable_hcs_processing context: {context}")
+    return bool(UNLEASH_CLIENT.is_enabled("hcs-data-processor", context))
 
 
 @celery_app.task(name="hcs.tasks.collect_hcs_report_data", queue=HCS_QUEUE)
